refactor(whatsapp_scanner): migrate to in-process CDP transport
Drop the local duplicated CdpConn / browser_ws_url / CDP_HOST / CDP_PORT
plumbing and route through crate::cdp::connect_and_attach_matching_in_process,
which uses the per-account in-process DevTools channel installed by
webview_accounts::open. Net effect: no TCP loopback attach needed for
either full or fast scan tick.

- Drop tokio_tungstenite + futures_util::{SinkExt, StreamExt} + connect_async
  imports (no remaining users in this module).
- Switch dom_snapshot.rs and idb.rs to crate::cdp::CdpConn.
- scan_dom_once now takes &AppHandle<R> so the in-process transport can be
  looked up; spawn_scanner loop passes it through.
- Predicate compares page url prefix/suffix; canonical helper already
  filters kind == "page".
